Revert "Merge branch 'kyle_cicd'"
This reverts commit 5bead5964d2c612f69bb1235f63eaf353f3fd1e8

+++ Backend/ms_313-main-Backend-debt_buddies_backend/Backend/debt_buddies_backend/springboot_application/src/main/java/debtbuddies/GameServer/PlayerClasses/User.java
@@ -6,14 +6,14 @@ public class User {
 
     protected final String username;
 
-    protected long playerID;
+    protected int playerID;
 
     protected int balance;
 
     public User(){
         playerID = playerID_index++;
         balance = 100;
-        username = Long.toString(playerID);
+        username = Integer.toString(playerID);
     }
 
     public User(String username){
@@ -22,7 +22,7 @@ public class User {
         this.username = username;
     }
 
-    public User(String username, long id, int balance){
+    public User(String username, int id, int balance){
         this.username = username;
         playerID = id;
         this.balance = balance;
@@ -36,7 +36,7 @@ public class User {
     
     public int getBalance(){ return balance; }
 
-    public long getID(){
+    public int getID(){
         return playerID;
     }
